[Ref] Add mpt::clamp, compatible with C++17 std::clamp. Convert common/, sounddev/ and sounddsp/.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@7225 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ sounddev/SoundDeviceWaveout.h
@@ -47,7 +47,7 @@ protected:
 	HANDLE m_ThreadWakeupEvent;
 	bool m_Failed;
 	HWAVEOUT m_hWaveOut;
-	ULONG m_nWaveBufferSize;
+	uint32 m_nWaveBufferSize;
 	bool m_JustStarted;
 	ULONG m_nPreparedHeaders;
 	ULONG m_nWriteBuffer;
